Fix sounds on Linux and Mac OS
README.md:
pyproject.toml:
requirements.txt:
  Upated modules needed for sound on non-Windows systems

changes/181.bugfix:
  Describe this fix for future release

d_rats/dplatform*.py:
docs/source/d_rats.rst:
  Move each platform into its own module.
  Add generic sound support.
