+++ changes/191.bugfix
@@ -1,0 +1,1 @@
+Fixed sound on Linux and enabled it for Mac OS.
